set separate stateless component for the API
