+++ src/Redux/home/homeSlice.js
@@ -1,23 +1,5 @@
-import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
-import axios from 'axios';
-
-const AQI_URL = 'https://api.waqi.info/feed/';
-const token = '/?token=84ae272c2a6c1c6890572f398be954a05210aaed';
-
-const cleanData = (apiData) => {
-  const {
-    aqi, idx, attributions, time, city,
-  } = apiData.data;
-  return {
-    aqi, idx, attributions, time, city,
-  };
-};
-
-export const fetchAqi = createAsyncThunk('aqi/fetchAqi', async (city) => {
-  const { data } = await (axios.get(`${AQI_URL}${city}${token}`));
-  const citiesData = cleanData(data);
-  return { ...citiesData, status: 'success' };
-});
+import { createSlice } from '@reduxjs/toolkit';
+import fetchAqi from '../apiFilter';
 
 const initialState = [];
